Make shape requests interruptible. (#65)
When building shape params on the server, for example using
`sync_render`, it's useful to be able to respond to database changes
that would cause the shape params to be different.

For example, `/sync/projects` -> looks up posts for user, syncs `where
id IN $userProjectIds`. Then the user joins another thread. Now, ... at
the moment, this can wait until 20s for a long poll to return before the
/sync/projects controller code runs to reconstruct the new shape params
with the new where clause.

What this PR does is provide a
`Phoenix.Sync.interrupt(Projects.Project)` function that other backend
code can call in order to interrupt current long polling requests. This
is similar to using the abort controller with the javascript client.

---------

Co-authored-by: Garry Hill <garry@magnetised.net>